Add support to delete Objective

+++ src/app/shared/components/deleting-item/deleting-item.component.ts
@@ -3,6 +3,7 @@ import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { Observable } from 'rxjs';
 import { catchError, map } from 'rxjs/operators';
 import { DeleteItemService } from 'src/app/core/services/delete-item.service';
+import { ActivityTrackerService } from 'src/app/pages/standard-reports/services/activity-tracker.service';
 
 @Component({
   selector: 'app-deleting-item',
@@ -16,21 +17,35 @@ export class DeletingItemComponent implements OnInit {
   deletedItem: boolean = false;
   deletingItem: boolean = false;
   deleteResponse$: Observable<any>;
+  data: any;
+  dialogCloseData: any;
   constructor(
     private dialogRef: MatDialogRef<DeletingItemComponent>,
     @Inject(MAT_DIALOG_DATA) data,
-    private deleteItemService: DeleteItemService
+    private deleteItemService: DeleteItemService,
+    private activityTrackerService: ActivityTrackerService
   ) {
     this.deletePath = data?.path;
     this.itemName = data?.itemName;
+    this.data = data;
+    this.deletedItem = false;
   }
 
   ngOnInit(): void {}
 
-  onConfirmDelete(e) {
+  onConfirmDelete(e: Event): void {
     e.stopPropagation();
     this.deletingItem = true;
-    this.deleteResponse$ = this.deleteItemService.deleteItem(this.deletePath);
+    if (!this.data?.dataStore) {
+      this.deleteResponse$ = this.deleteItemService.deleteItem(this.deletePath);
+    } else {
+      this.deleteResponse$ =
+        this.activityTrackerService.saveActivityTrackerDetails(
+          this.data?.path,
+          this.data?.data
+        );
+    }
+
     this.deleteResponse$.subscribe((response) => {
       if (response) {
         setTimeout(() => {
